Update to 1.21.5

+++ src/main/kotlin/io/github/dockyardmc/player/Player.kt
@@ -20,7 +20,6 @@ import io.github.dockyardmc.extentions.sendPacket
 import io.github.dockyardmc.inventory.ContainerInventory
 import io.github.dockyardmc.inventory.PlayerInventory
 import io.github.dockyardmc.inventory.give
-import io.github.dockyardmc.item.EquipmentSlot
 import io.github.dockyardmc.item.ItemStack
 import io.github.dockyardmc.location.Location
 import io.github.dockyardmc.maths.percentOf
@@ -36,6 +35,7 @@ import io.github.dockyardmc.protocol.packets.ProtocolState
 import io.github.dockyardmc.protocol.packets.play.clientbound.*
 import io.github.dockyardmc.protocol.packets.play.serverbound.ServerboundChatCommandPacket
 import io.github.dockyardmc.protocol.packets.play.serverbound.ServerboundClientInputPacket
+import io.github.dockyardmc.protocol.types.EquipmentSlot
 import io.github.dockyardmc.registry.Blocks
 import io.github.dockyardmc.registry.EntityTypes
 import io.github.dockyardmc.registry.Items
@@ -457,7 +457,7 @@ class Player(
         sendPacket(packet)
     }
 
-    fun playTotemAnimation(customModelData: Int? = null) {
+    fun playTotemAnimation(customModelData: Float? = null) {
         val held = getHeldItem(PlayerHand.MAIN_HAND)
         if (customModelData != null) {
             val totem = ItemStack(Items.TOTEM_OF_UNDYING).withCustomModelData(customModelData)
